Implementation of optimising media library uploads with TinyPNG
To make this work the upload route defined within `OrchardCore.Media`
is overridden using a `Route` attribute. Most of the contents of the
action is copied from `OrchardCore.Media` accept for the call to create
the file in the media store. The `Stream` parameter is passed to a
`TinyPngService` that'll will use `Tinify` to optimise the image via
TinyPNG. In order to optimise images with TinyPNG there must be an API
key from a developer account, which can be managed via site settings.
If no API key is present or the image isn't a `.jpg` or `.png` then the
file won't be optimised.